Use FieldWriteback
